Policies & Roles

+++ app/Http/Controllers/AppointmentController.php
@@ -28,16 +28,22 @@ class AppointmentController extends Controller
 
     public function index()
     {
+        $this->authorize('viewAny', Appointment::class);
+
         return new AppointmentCollection(Appointment::paginate(10));
     }
 
     public function show(Appointment $appointment)
     {
+        $this->authorize('view', $appointment);
+
         return response()->json(new AppointmentResource($appointment), 200);
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Appointment::class);
+
         $request->validate(self::$rules, self::$messages);
         $appointment = new Appointment($request->all());
         $appointment->save();
@@ -47,6 +53,8 @@ class AppointmentController extends Controller
 
     public function update(Request $request, Appointment $appointment)
     {
+        $this->authorize('update', $appointment);
+
         $request->validate(self::$repulses, self::$messages);
         $appointment->update($request->all());
         return response()->json($appointment, 200);
@@ -54,6 +62,8 @@ class AppointmentController extends Controller
 
     public function delete(Request $request, Appointment $appointment)
     {
+        $this->authorize('delete', $appointment);
+
         $appointment->delete();
         return response()->json(null, 204);
     }
